Set up SQLDelight with caching

+++ shared/src/commonMain/kotlin/name/patrikspano/mobilecasestudy/entity/Entity.kt
@@ -31,7 +31,7 @@ data class SearchCoin(
     val apiSymbol: String,
     val symbol: String,
     @SerialName("market_cap_rank")
-    val marketCapRank: Int,
+    val marketCapRank: Long,
     val thumb: String,
     val large: String
 )
@@ -40,16 +40,16 @@ data class SearchCoin(
 data class TrendingCoin(
     val id: String,
     @SerialName("coin_id")
-    val coinId: Int,
+    val coinId: Long,
     val name: String,
     val symbol: String,
     @SerialName("market_cap_rank")
-    val marketCapRank: Int,
+    val marketCapRank: Long,
     val thumb: String,
     val small: String,
     val large: String,
     val slug: String,
     @SerialName("price_btc")
     val priceBtc: Double,
-    val score: Int
+    val score: Long
 )
